Creating new file for handle_updates and removing functions of file taskbot
